Partially use native types for storage core

+++ src/alias.c
@@ -27,7 +27,7 @@ struct _alias_list {
 } __attribute__((packed));
 
 /* Read list structure from offset */
-static struct _alias_list *get_alias_list(base_t *base, unsigned long long offset) {
+static struct _alias_list *get_alias_list(base_t *base, uint64_t offset) {
 	int fd = pager_get_fd(base, &offset);
 
 	struct _alias_list *list = (struct _alias_list *)zcalloc(1, sizeof(struct _alias_list));
@@ -49,7 +49,7 @@ static struct _alias_list *get_alias_list(base_t *base, unsigned long long offse
 	return list;
 }
 
-static void flush_alias_list(base_t *base, struct _alias_list *list, unsigned long long offset) {
+static void flush_alias_list(base_t *base, struct _alias_list *list, uint64_t offset) {
 	int fd = pager_get_fd(base, &offset);
 
 	if (lseek(fd, offset, SEEK_SET) < 0) {
